refactor(backfill): share backfillBustouts core between callable and script
Extract the #214 `backfillBustoutsForShows` pipeline into
`functions/backfillBustoutsCore.js` and the server scoring helpers into
`functions/scoringCore.js`. The deployed callable and the admin ops script
now both import the same `runBackfill(...)`, so the script no longer has
to round-trip through Firebase Auth (custom token → Identity Toolkit
`signInWithCustomToken` → Bearer to the HTTPS callable) just to reach
code it could call directly.

Why: running the script locally against prod kept hitting every Firebase
auth layer in turn — `invalid_rapt`, `iam.serviceAccounts.signBlob`
denied, and `Requests from referer <empty> are blocked`. Each fix was a
separate IAM / API-key chore. Since the script already has admin-level
Firestore via ADC, the HTTPS round-trip was pure ceremony.

Behavior unchanged:
- Callable still gated by `assertAdminClaim`; input contract identical
  (`{ showDates }` or `{ mode: "missing" }`); return shape identical.
- Pipeline is still a single source of truth — sharing just happens at
  module-import time instead of over HTTPS.
- Script remains idempotent; re-running with unchanged Phish.net data
  produces zero score deltas.

New local workflow:
  gcloud auth application-default login
  cd functions
  PHISHNET_API_KEY="$(firebase functions:secrets:access PHISHNET_API_KEY \
    --project=<project-id>)" \
    npm run backfill:bustouts -- --showDates=YYYY-MM-DD --apply

No more SA impersonation, `iamcredentials` API enablement, Token Creator
binding, or unrestricted admin API key.

Tests: 85/85 functions node:test + 40/40 client vitest passing.
Made-with: Cursor
