+++ functions/index.js
@@ -11,12 +11,8 @@ const {
   syncPhishnetSongCatalogToStorage,
 } = require("./phishnetSongCatalog");
 const {
-  BUSTOUT_MIN_GAP: AUTOMATION_BUSTOUT_MIN_GAP,
   candidateShowDates,
-  deriveBustoutsFromRows,
-  fetchPhishnetSetlistForDate,
   isWithinLiveSetlistPollWindow,
-  normalizeSetlistRows,
   parseShowCalendarSnapshotToDateSet,
   pollSingleShowDate,
   scheduledCandidateShowDates,
@@ -31,141 +27,16 @@ const {
   findPoolPickActivity,
   parseShowCalendarDates,
 } = require("./poolDelete");
+const {
+  calculateTotalScore,
+  actualSetlistFromOfficialDoc,
+} = require("./scoringCore");
+const { runBackfill } = require("./backfillBustoutsCore");
 
 const phishnetApiKey = defineSecret("PHISHNET_API_KEY");
 
 admin.initializeApp();
 const db = admin.firestore();
-
-// Mirrors src/utils/scoring.js (keep in sync).
-const SCORING_RULES = {
-  EXACT_SLOT: 10,
-  ENCORE_EXACT: 15,
-  IN_SETLIST: 5,
-  WILDCARD_HIT: 10,
-  BUSTOUT_BOOST: 20,
-  BUSTOUT_MIN_GAP: 30,
-};
-
-const SCORE_FIELDS = ["s1o", "s1c", "s2o", "s2c", "enc", "wild"];
-
-/** Keys on the scoring payload that are not slot song strings. */
-const NON_SONG_SETLIST_KEYS = new Set([
-  "officialSetlist",
-  "encoreSongs",
-  "id",
-  "bustouts",
-]);
-
-function normalizeSong(value) {
-  return String(value ?? "").trim().toLowerCase();
-}
-
-function guessMatchesEncoreExact(actualSetlist, guessNorm) {
-  if (!guessNorm) return false;
-  const primary = normalizeSong(actualSetlist.enc);
-  if (primary === guessNorm) return true;
-  const list = actualSetlist.encoreSongs;
-  if (!Array.isArray(list)) return false;
-  return list.some((t) => normalizeSong(t) === guessNorm);
-}
-
-/**
- * Slot strings from the flat map plus ordered officialSetlist, normalized and deduped.
- * Same as buildAllPlayedNormalized in src/utils/scoring.js.
- * @param {Record<string, unknown>} actualSetlist
- * @returns {string[]}
- */
-function buildAllPlayedNormalized(actualSetlist) {
-  const fromSlots = [];
-  for (const [key, val] of Object.entries(actualSetlist || {})) {
-    if (NON_SONG_SETLIST_KEYS.has(key)) continue;
-    if (typeof val !== "string") continue;
-    const t = val.trim();
-    if (!t) continue;
-    fromSlots.push(normalizeSong(t));
-  }
-
-  const rawOfficial = actualSetlist.officialSetlist;
-  const fromOfficial = Array.isArray(rawOfficial)
-    ? rawOfficial
-        .map((s) =>
-          typeof s === "string" ? s.trim() : String(s ?? "").trim()
-        )
-        .filter(Boolean)
-        .map((s) => normalizeSong(s))
-    : [];
-
-  const combined = [...fromSlots, ...fromOfficial];
-  return [...new Set(combined)];
-}
-
-/**
- * Mirrors computeSlotResult + bustout from getSlotScoreBreakdown in
- * src/shared/utils/scoring.js.
- *
- * Bustout boosts read the per-show `bustouts` snapshot on the official
- * setlist doc (#214). The snapshot is frozen at save time from Phish.net row
- * `gap`, so scoring is deterministic and never drifts with the weekly
- * `song-catalog.json` refresh. Absence / empty array → no bustout boost; no
- * catalog fallback.
- */
-function calculateSlotScore(fieldId, guessedSong, actualSetlist) {
-  if (!actualSetlist || !guessedSong) return 0;
-
-  const guess = normalizeSong(guessedSong);
-  if (!guess) return 0;
-
-  const allPlayed = buildAllPlayedNormalized(actualSetlist);
-
-  let base = 0;
-  if (fieldId === "wild") {
-    if (allPlayed.includes(guess)) {
-      base = SCORING_RULES.WILDCARD_HIT;
-    } else {
-      return 0;
-    }
-  } else {
-    const exactNonEnc =
-      fieldId !== "enc" && normalizeSong(actualSetlist[fieldId]) === guess;
-    const exactEnc =
-      fieldId === "enc" && guessMatchesEncoreExact(actualSetlist, guess);
-    if (exactNonEnc || exactEnc) {
-      base =
-        fieldId === "enc"
-          ? SCORING_RULES.ENCORE_EXACT
-          : SCORING_RULES.EXACT_SLOT;
-    } else if (allPlayed.includes(guess)) {
-      base = SCORING_RULES.IN_SETLIST;
-    } else {
-      return 0;
-    }
-  }
-
-  const bustoutList = Array.isArray(actualSetlist.bustouts)
-    ? actualSetlist.bustouts
-    : [];
-  let bustoutBoost = false;
-  for (const raw of bustoutList) {
-    if (typeof raw !== "string") continue;
-    if (normalizeSong(raw) === guess) {
-      bustoutBoost = true;
-      break;
-    }
-  }
-
-  return base + (bustoutBoost ? SCORING_RULES.BUSTOUT_BOOST : 0);
-}
-
-function calculateTotalScore(userPicks, actualSetlist) {
-  if (!actualSetlist || !userPicks) return 0;
-  return SCORE_FIELDS.reduce((total, fieldId) => {
-    return (
-      total +
-      calculateSlotScore(fieldId, userPicks[fieldId], actualSetlist)
-    );
-  }, 0);
-}
 
 /**
  * Gate an admin-only callable on the `admin: true` custom claim (issue #139).
@@ -243,24 +114,6 @@ async function recomputeLiveScoresForShow(showDate, actualSetlistFromWrite = nul
     await batch.commit();
   }
   return { updatedPicks };
-}
-
-function actualSetlistFromOfficialDoc(setlistDoc) {
-  const setlistFlat = setlistDoc.setlist || {};
-  const out = {
-    ...setlistFlat,
-    officialSetlist: Array.isArray(setlistDoc.officialSetlist)
-      ? setlistDoc.officialSetlist
-      : [],
-  };
-  if (Array.isArray(setlistDoc.encoreSongs) && setlistDoc.encoreSongs.length > 0) {
-    out.encoreSongs = setlistDoc.encoreSongs;
-  }
-  // Per-show bustout snapshot for scoring (#214). Absence/empty → no boost.
-  if (Array.isArray(setlistDoc.bustouts)) {
-    out.bustouts = setlistDoc.bustouts;
-  }
-  return out;
 }
 
 exports.gradePicksOnSetlistWrite = onDocumentWritten(
@@ -1015,136 +868,39 @@ exports.backfillBustoutsForShows = onCall(
       );
     }
 
-    // Resolve target show dates.
-    /** @type {string[]} */
-    let showDates = [];
     const explicitDates = request.data?.showDates;
-    if (Array.isArray(explicitDates) && explicitDates.length > 0) {
-      showDates = explicitDates.map((d) => assertShowDateString(d));
-    } else if (request.data?.mode === "missing") {
-      const snap = await db.collection("official_setlists").get();
-      for (const d of snap.docs) {
-        const data = d.data() || {};
-        if (!Array.isArray(data.bustouts)) {
-          showDates.push(d.id);
-        }
-      }
-    } else {
+    const mode = request.data?.mode === "missing" ? "missing" : undefined;
+    if (
+      !(Array.isArray(explicitDates) && explicitDates.length > 0) &&
+      mode !== "missing"
+    ) {
       throw new HttpsError(
         "invalid-argument",
         'Pass { showDates: ["YYYY-MM-DD", ...] } or { mode: "missing" }.'
       );
     }
 
-    const results = [];
-    for (const showDate of showDates) {
-      const setlistRef = db.collection("official_setlists").doc(showDate);
-      const setlistSnap = await setlistRef.get();
-      if (!setlistSnap.exists) {
-        results.push({ showDate, skipped: "no-setlist-doc" });
-        continue;
-      }
-
-      // 1) Re-fetch Phish.net rows and derive bustouts.
-      let bustouts;
-      try {
-        const payload = await fetchPhishnetSetlistForDate(
-          showDate,
-          String(key).trim()
-        );
-        const rows = normalizeSetlistRows(payload);
-        bustouts = deriveBustoutsFromRows(rows);
-      } catch (e) {
-        const msg = e instanceof Error ? e.message : String(e);
-        logger.warn("backfillBustoutsForShows fetch failed", { showDate, msg });
-        results.push({ showDate, skipped: "phishnet-fetch-failed", error: msg });
-        continue;
-      }
-
-      // 2) Write merge + capture prior bustouts for idempotency / logging.
-      const prior = setlistSnap.data() || {};
-      const priorBustouts = Array.isArray(prior.bustouts) ? prior.bustouts : null;
-      await setlistRef.set(
-        {
-          bustouts,
-          updatedAt: admin.firestore.FieldValue.serverTimestamp(),
-          updatedBy: request.auth?.token?.email || "backfill-bustouts",
-        },
-        { merge: true }
-      );
-
-      // 3) Recompute live scores from the *post-write* doc so the snapshot
-      // we just wrote is what scoring sees. Also reconcile graded picks'
-      // contribution to `users.totalPoints` by score diff — mirrors the
-      // rollup pathway so we don't leave stale totals behind.
-      const freshSnap = await setlistRef.get();
-      const freshDoc = freshSnap.data() || {};
-      const actualSetlist = actualSetlistFromOfficialDoc(freshDoc);
-
-      const picksSnap = await db
-        .collection("picks")
-        .where("showDate", "==", showDate)
-        .get();
-
-      let batch = db.batch();
-      let opCount = 0;
-      let updatedPicks = 0;
-      let reconciledGradedPicks = 0;
-
-      for (const pickDoc of picksSnap.docs) {
-        if (opCount + 2 > MAX_FIRESTORE_BATCH_WRITES) {
-          await batch.commit();
-          batch = db.batch();
-          opCount = 0;
-        }
-        const pickData = pickDoc.data() || {};
-        const newScore = calculateTotalScore(pickData.picks || {}, actualSetlist);
-        const oldScore = Number.isFinite(pickData.score) ? Number(pickData.score) : 0;
-        const scoreDelta = newScore - oldScore;
-
-        const pickUpdate = { score: newScore };
-        if (pickData.isGraded !== true) {
-          pickUpdate.gradedAt = admin.firestore.FieldValue.delete();
-        }
-        batch.update(pickDoc.ref, pickUpdate);
-        opCount += 1;
-        updatedPicks += 1;
-
-        // Only reconcile user totals when this pick was already graded —
-        // otherwise the rollup flow owns first-grade accounting.
-        if (pickData.isGraded === true && scoreDelta !== 0 && pickData.userId) {
-          batch.set(
-            db.collection("users").doc(pickData.userId),
-            {
-              totalPoints: admin.firestore.FieldValue.increment(scoreDelta),
-            },
-            { merge: true }
-          );
-          opCount += 1;
-          reconciledGradedPicks += 1;
-        }
-      }
-
-      if (opCount > 0) {
-        await batch.commit();
-      }
-
-      results.push({
-        showDate,
-        bustoutCount: bustouts.length,
-        priorBustoutCount: priorBustouts ? priorBustouts.length : null,
-        updatedPicks,
-        reconciledGradedPicks,
+    try {
+      const { results } = await runBackfill({
+        db,
+        admin,
+        logger,
+        phishnetApiKey: key,
+        showDates: Array.isArray(explicitDates) ? explicitDates : undefined,
+        mode,
+        updatedBy: request.auth?.token?.email || "backfill-bustouts",
       });
+      logger.info("backfillBustoutsForShows callable complete", {
+        callerUid: request.auth?.uid || null,
+      });
+      return { ok: true, results };
+    } catch (e) {
+      const msg = e instanceof Error ? e.message : String(e);
+      if (/^showDate must be a YYYY-MM-DD/.test(msg)) {
+        throw new HttpsError("invalid-argument", msg);
+      }
+      throw e;
     }
-
-    logger.info("backfillBustoutsForShows complete", {
-      minGap: AUTOMATION_BUSTOUT_MIN_GAP,
-      results,
-      callerUid: request.auth?.uid || null,
-    });
-
-    return { ok: true, results };
   }
 );
